requirements.txt
python-chess>=1.999
pytest
torch>=2.0
numpy
playwright
